fix: [ast-util] addCustomScalarType() 함수 옮김
