+++ lib/utilities/ast-filter.ts
@@ -1,4 +1,4 @@
-import { DocumentNode, Kind, ScalarTypeDefinitionNode, visit } from 'graphql'
+import { DocumentNode, Kind, visit } from 'graphql'
 
 export const filterOnlyVisitedSchema = ({
   ast,
@@ -62,33 +62,3 @@ export const filterOnlyVisitedSchema = ({
     }
   });
 };
-
-/**
- * 스키마에 custom scalar type을 추가합니다.
- * */
-export const addCustomScalarType = ({
-  ast,
-  customScalarName
-}: {
-  ast: DocumentNode,
-  customScalarName?: string
-}) => {
-  if (!customScalarName) {
-    return ast;
-  }
-
-  const scalarTypeDefinitionNode: ScalarTypeDefinitionNode = {
-    kind: Kind.SCALAR_TYPE_DEFINITION,
-    name: {
-      kind: Kind.NAME,
-      value: customScalarName,
-    },
-  };
-
-  const newAstDefinitions = [...ast.definitions, scalarTypeDefinitionNode]
-
-  return {
-    ...ast,
-    definitions: newAstDefinitions,
-  };
-}
